separa os arquivos de pictureDay
